- Merge branch 'master' of https://github.com/SupSuper/OpenXcom
into oXc_kL

Conflicts:
	bin/data/Language/en-GB.yml
	src/Basescape/MiniBaseView.cpp
	src/Basescape/SellState.h
	src/Basescape/SoldierMemorialState.cpp
	src/Basescape/TransferItemsState.h
	src/Battlescape/BattlescapeGenerator.cpp
	src/Battlescape/BattlescapeState.cpp
	src/Battlescape/BriefingState.cpp
	src/Battlescape/DebriefingState.cpp
	src/Battlescape/ExplosionBState.cpp
	src/Engine/Game.cpp
	src/Engine/Game.h
	src/Geoscape/ConfirmLandingState.cpp
	src/Geoscape/ConfirmLandingState.h
	src/Geoscape/DefeatState.cpp
	src/Geoscape/GeoscapeState.cpp
	src/Geoscape/GeoscapeState.h
	src/Geoscape/MonthlyReportState.cpp
	src/Geoscape/PsiTrainingState.cpp
	src/Geoscape/PsiTrainingState.h
	src/Geoscape/VictoryState.cpp
	src/Interface/TextList.cpp
	src/Menu/MainMenuState.cpp
	src/Menu/NewBattleState.cpp
	src/Menu/NewBattleState.h
	src/Menu/OptionsAudioState.cpp
	src/Menu/OptionsAudioState.h
	src/Menu/OptionsBaseState.cpp
	src/OpenXcom.2010.sln
	src/Resource/ResourcePack.cpp
	src/Resource/ResourcePack.h
	src/Ruleset/AlienDeployment.cpp
	src/Ruleset/AlienDeployment.h
	src/Ruleset/RuleAlienMission.h
	src/Ruleset/RuleTerrain.cpp
	src/Ruleset/RuleTerrain.h
	src/Ruleset/UfoTrajectory.h
	src/main.cpp
	src/pch.h

+++ src/Basescape/MiniBaseView.cpp
@@ -101,7 +101,7 @@ size_t MiniBaseView::getHoveredBase() const
  * Changes the base that is currently selected on the mini base view.
  * @param base ID of base.
  */
-void MiniBaseView::setSelectedBase(unsigned base)
+void MiniBaseView::setSelectedBase(size_t base)
 {
 	_base = base;
