+++ src/Basescape/SellState.cpp
@@ -335,7 +335,7 @@ void SellState::think()
  * @param selected Selected craft.
  * @return Index of the selected craft.
  */
-int SellState::getCraftIndex(unsigned selected) const
+int SellState::getCraftIndex(size_t selected) const
 {
 	return selected - _soldiers.size();
 }
@@ -779,9 +779,9 @@ void SellState::updateItemStrings()
  * @param selected Currently selected item.
  * @return The type of the selected item.
  */
-enum SellType SellState::getType(unsigned selected) const
+enum SellType SellState::getType(size_t selected) const
 {
-	unsigned max = _soldiers.size();
+	size_t max = _soldiers.size();
 
 	if (selected < max)
 		return SELL_SOLDIER;
@@ -803,7 +803,7 @@ enum SellType SellState::getType(unsigned selected) const
  * @param selected Currently selected item.
  * @return Index of the selected item.
  */
-int SellState::getItemIndex(unsigned selected) const
+int SellState::getItemIndex(size_t selected) const
 {
 	return static_cast<int>(selected)
 			- static_cast<int>(_soldiers.size())
