+++ src/Basescape/TransferItemsState.h
@@ -98,9 +98,9 @@ private:
 	/// Gets distance between bases.
 	double getDistance() const;
 	/// Gets type of selected item.
-	enum TransferType getType(unsigned selected) const;
+	enum TransferType getType(size_t selected) const;
 	/// Gets item Index.
-	int getItemIndex(unsigned selected) const;
+	int getItemIndex(size_t selected) const;
 
 
 	public:
